update openapi test structure
